+++ js/boot-ver.js
@@ -1,2 +1,2 @@
 'use strict';
-window.VER = '5.1.19';
+window.VER = '5.1.21';